Styles and resource dictionaries
Added the part of the pull request for ProgressIndicator, so now it
stretches
correctly on start of the 'busy'. Still doesn't stretch on resize.

Added few styles.

TODO: Fix the resource dictionary reference in InputGenerationView, now
the application
either crashes when the resource dictionary is included or the view does
not
find the references to styles.